refactor: King과 Guard의 중복된 코드를 중간 추상 클래스로 분리
